code/requirements.txt
argparse
matplotlib
numpy
Pillow
tensorflow
scikit-learn
